Change int to unsigned long long int for max: HW 2

+++ fib.c
@@ -1,13 +1,13 @@
 #include <stdio.h>
 #include <stdlib.h>
 
-int FibIterative(int n) {
+unsigned long long int FibIterative(int n) {
    if (n == 1) {
       return 0;
    } else if (n == 2) {
       return 1;
    }
-   int first = 0, second = 1, next_value;
+   unsigned long long int first = 0, second = 1, next_value;
    for (int i = 3; i <= n; i++) {
       next_value = first + second;
       first = second;
@@ -16,7 +16,7 @@ int FibIterative(int n) {
    return second;
 }
 
-int FibRecursive(int n) {
+unsigned long long int FibRecursive(int n) {
    if(n == 1){
       return 0;
    } else if (n == 2) {
@@ -37,11 +37,11 @@ int main(int argc, char *argv[]) {
    int N = value + file_value;
 
    if (argv[2][0] == 'i') {
-      int result = FibIterative(N);
-      printf("%d", result);
+     unsigned long long int result = FibIterative(N);
+      printf("%llu", result);
    } else if (argv[2][0] == 'r') {
-      int result = FibRecursive(N);
-      printf("%d", result);
+      unsigned long long int result = FibRecursive(N);
+      printf("%llu", result);
    }
    return 0;
